chore: rename module to core
