+++ game/tests/unit/test_mines.py
@@ -1,5 +1,7 @@
 import pytest
-from game.src.mechanics.mines import randomly_generate_mines, precompute_adjacent_mines
+
+from game.src.core.mines import precompute_adjacent_mines, randomly_generate_mines
+
 
 def should_randomly_generate_mines_positions_given_rows_cols_and_a_number_of_expected_mines():
     expected_positions = {(3, 8), (4, 3), (0, 0), (1, 8), (1, 0), (1, 6), (3, 2), (6, 3), (1, 3), (0, 8)}
